+++ demos/Betty/Zombie.py
@@ -5,7 +5,7 @@ from lge.Engine import Engine
 class Zombie( Sprite ):
     def __init__( self, name ):
         super().__init__( "zombie", (0,0), name  )
-        self.SetShape( 0, "zombie" )
+        self.SetShape( "zombie", 0 )
         self.tag = "zombie"
         self.dir = "R"
 
@@ -85,7 +85,7 @@ class Zombie( Sprite ):
 
         # tunel?
         x, y = self.GetPosition()
-        w, h = Engine.GetWorldBounds().GetSize()
+        w, h = Engine.GetCamera().GetSize()
         if( x < -16 ): x = w - 16
         elif( x > w - 16 ): x = -16
         self.SetPosition( (x,y) )