+++ demos/OOP/Demo02_MoveCamera.py
@@ -8,8 +8,7 @@ class MiJuego():
     def __init__( self ):
         # creamos el juego
         Engine.Init( (640,480), "Move Camera" )
-        Engine.SetWorldBounds( Rect( (0,0), (1920,1056) ) )
-        Engine.SetMainTask( self.MainControl )
+        Engine.SetUpdate( self.MainUpdate )
 
         # activamos la musica de fondo
         Engine.LoadSound( "fondo", "../sounds/happy-and-sad.wav" )
@@ -33,13 +32,17 @@ class MiJuego():
         infobar = Text( None, (0,460), "monospace", (0,0,0), None, "infobar" )
         Engine.AddGObject( infobar, Engine.CAM_LAYER )
 
+        # configuramos la camara
+        camera = Engine.GetCamera()
+        camera.SetBounds( Rect( (0,0), (1920,1056) ) )
+
         # posicionamos la camara
         x, y = heroe.GetPosition()
         w, h = heroe.GetSize()
-        cw, ch = Engine.GetCamSize()
-        Engine.SetCamPosition( (x+w/2-cw/2,y+h/2-ch/2) )
+        cw, ch = camera.GetSize()
+        camera.SetPosition( (x+w/2-cw/2,y+h/2-ch/2) )
 
-    def MainControl( self, dt ):
+    def MainUpdate( self, dt ):
         # abortamos con la tecla Escape
         if( Engine.IsKeyPressed( Engine.CONSTANTS.K_ESCAPE ) ):
             Engine.Quit()
@@ -60,7 +63,8 @@ class MiJuego():
         pixels = round( (pps*dt)/1000 )
 
         # la posiciona actual de la camara
-        x, y = Engine.GetCamPosition()
+        camera = Engine.GetCamera()
+        x, y = camera.GetPosition()
 
         # cambiamos sus coordenadas segun la tecla presionada
         if( Engine.IsKeyPressed( Engine.CONSTANTS.K_RIGHT ) ):
@@ -73,7 +77,7 @@ class MiJuego():
             y = y + pixels
 
         # posicionamos la camara
-        Engine.SetCamPosition( (x,y) )
+        camera.SetPosition( (x,y) )
 
 
     # main loop