+++ demos/OOP/Demo04_AnimatedPlayer.py
@@ -7,8 +7,7 @@ class MiJuego():
     def __init__( self ):
         # creamos el juego
         Engine.Init( (640,480), "Animated Player" )
-        Engine.SetWorldBounds( Rect( (0,0), (1920,1056) ) )
-        Engine.SetMainTask( self.MainControl )
+        Engine.SetUpdate( self.MainUpdate )
 
         # activamos la musica de fondo
         Engine.LoadSound( "fondo", "../sounds/happy-and-sad.wav" )
@@ -32,10 +31,14 @@ class MiJuego():
         infobar = Text( None, (0,460), "monospace", (0,0,0), None, "infobar" )
         Engine.AddGObject( infobar, Engine.CAM_LAYER )
 
-        # establecemos que la camara siga al heroe en su origen
-        Engine.SetCamTarget( heroe, False )
+        # configuramos la camara
+        camera = Engine.GetCamera()
+        camera.SetBounds( Rect( (0,0), (1920,1056) ) )
 
-    def MainControl( self, dt ):
+        # establecemos que la camara siga al heroe
+        Engine.SetCameraTarget( heroe, True )
+
+    def MainUpdate( self, dt ):
         # abortamos con la tecla Escape
         if( Engine.IsKeyPressed( Engine.CONSTANTS.K_ESCAPE ) ):
             Engine.Quit()
@@ -61,7 +64,7 @@ class MiHeroe( Sprite ):
         # agregamos el heroe con diferentes imagenes
         super().__init__( ["idle","run"], (550,346), "Heroe" )
         self.Scale( 0.16 )
-        self.SetShape( 0, "idle" )
+        self.SetShape( "idle", 0 )
         self.heading = 1
 
     def OnUpdate( self, dt ):
@@ -74,22 +77,22 @@ class MiHeroe( Sprite ):
 
         # cambiamos sus coordenadas, orientacion e imagen segun la tecla presionada
         moving = False
-        idx, name = self.GetCurrentShape()
+        iname, idx = self.GetCurrentShape()
         if( Engine.IsKeyPressed( Engine.CONSTANTS.K_RIGHT ) ):
             x = x + pixels
             if( self.heading != 1 ):
                 self.Flip( True, False )
                 self.heading = 1
-            if( name != "run" ):
-                self.SetShape( 0, "run" )
+            if( iname != "run" ):
+                self.SetShape( "run", 0 )
             moving = True
         elif( Engine.IsKeyPressed( Engine.CONSTANTS.K_LEFT ) ):
             x = x - pixels
             if( self.heading != -1 ):
                 self.Flip( True, False )
                 self.heading = -1
-            if( name != "run" ):
-                self.SetShape( 0, "run" )
+            if( iname != "run" ):
+                self.SetShape( "run", 0 )
             moving = True
 
         if( Engine.IsKeyPressed( Engine.CONSTANTS.K_DOWN ) ):
@@ -99,14 +102,16 @@ class MiHeroe( Sprite ):
             y = y + pixels
             moving = True
 
-        if( not moving and name != "idle" ):
-                self.SetShape( 0, "idle" )
+        if( not moving and iname != "idle" ):
+                self.SetShape("idle", 0 )
 
         # siguiente imagen de la secuencia
         self.NextShape( dt, 50 )
 
         # lo posicionamos asegurando que se encuentre dentro del mundo definido
-        self.SetPosition( (x,y), Engine.GetWorldBounds() )
+        camera = Engine.GetCamera()
+        bounds = camera.GetBounds()
+        self.SetPosition( (x,y), bounds )
 
 
 #--- show time