+++ demos/Procedural/Demo01_TheWorld.py
@@ -2,9 +2,9 @@ from lge.Engine import Engine
 from lge.Sprite import Sprite
 from lge.Rect import Rect
 
+
 # creamos el juego
 Engine.Init( (800,440), "The World" )
-Engine.SetWorldBounds( Rect( (0,0), (800,440) ) )
 
 # activamos la musica de fondo
 Engine.LoadSound( "fondo", "../sounds/happy-and-sad.wav" )
@@ -23,9 +23,6 @@ Engine.AddGObject( fondo, 0 )
 heroe = Sprite( "heroe", (226,142), "Heroe" )
 heroe.Scale( 0.08 )
 Engine.AddGObject( heroe, 1 )
-
-# posicionamos la camara
-Engine.SetCamPosition( (0,0) )
 
 # python un poco mas avanzado
 heroe.OnUpdate = lambda dt: heroe.NextShape(dt,60)