+++ demos/Procedural/Demo02_MoveCamera.py
@@ -4,7 +4,7 @@ from lge.Text import Text
 from lge.Rect import Rect
 
 
-def MainControl( dt ):
+def MainUpdate( dt ):
     # abortamos con la tecla Escape
     if( Engine.IsKeyPressed( Engine.CONSTANTS.K_ESCAPE ) ):
         Engine.Quit()
@@ -25,7 +25,8 @@ def MainControl( dt ):
     pixels = round( (pps*dt)/1000 )
 
     # la posiciona actual de la camara
-    x, y = Engine.GetCamPosition()
+    camera = Engine.GetCamera()
+    x, y = camera.GetPosition()
 
     # cambiamos sus coordenadas segun la tecla presionada
     if( Engine.IsKeyPressed( Engine.CONSTANTS.K_RIGHT ) ):
@@ -38,14 +39,13 @@ def MainControl( dt ):
         y = y + pixels
 
     # posicionamos la camara
-    Engine.SetCamPosition( (x,y) )
+    camera.SetPosition( (x,y) )
 
 
 def main():
     # creamos el juego
     Engine.Init( (640,480), "Move Camera" )
-    Engine.SetWorldBounds( Rect( (0,0), (1920,1056) ) )
-    Engine.SetMainTask( MainControl )
+    Engine.SetUpdate( MainUpdate )
 
     # activamos la musica de fondo
     Engine.LoadSound( "fondo", "../sounds/happy-and-sad.wav" )
@@ -69,11 +69,15 @@ def main():
     infobar = Text( None, (0,460), "monospace", (0,0,0), None, "infobar" )
     Engine.AddGObject( infobar, Engine.CAM_LAYER )
 
+    # configuramos la camara
+    camera = Engine.GetCamera()
+    camera.SetBounds( Rect( (0,0), (1920,1056) ) )
+
     # posicionamos la camara
     x, y = heroe.GetPosition()
     w, h = heroe.GetSize()
-    cw, ch = Engine.GetCamSize()
-    Engine.SetCamPosition( (x+w/2-cw/2,y+h/2-ch/2) )
+    cw, ch = camera.GetSize()
+    camera.SetPosition( (x+w/2-cw/2,y+h/2-ch/2) )
 
     # main loop
     Engine.Run( 60 )